Save HDK shell's checkpoint in aws-fpga-hdk-resources bucket instead of repo.
- Added upload_shell_checkpoint.sh
- Download shell checkpoint in hdk_setup.sh.
- Check version of checkpoint in hkd_setup.sh.
- Uses sha1 sum of the checkpoint that is stored in the bucket.

+++ hdk_setup.sh
@@ -12,33 +12,41 @@ if [[ ':$AWS_FPGA_REPO_DIR' == ':' ]]; then
   export AWS_FPGA_REPO_DIR=$script_dir
 fi
 
-echo "AWS FPGA: hdk_setup.sh script will:"
-echo "AWS FPGA:   (i) check if Xilinx's vivado is installed,"
-echo "AWS FPGA:   (ii) set up key environment variables HDK_*, and"
-echo "AWS FPGA:   (iii) prepare DRAM controller and PCIe IP modules if they are not already available in your directory."
+function info_msg {
+  echo -e "AWS FPGA: $1"
+}
 
-echo "AWS FPGA: Checking for vivado install:"
- 
+function err_msg {
+  echo >&2 "AWS FPGA: ERROR - $1"
+}
+
+info_msg "hdk_setup.sh script will:"
+info_msg "  (i) check if Xilinx's vivado is installed,"
+info_msg "  (ii) set up key environment variables HDK_*, and"
+info_msg "  (iii) prepare DRAM controller and PCIe IP modules if they are not already available in your directory."
+
+info_msg "Checking for vivado install:"
+
 # before going too far make sure Vivado is available
-vivado -version >/dev/null 2>&1 || { echo >&2 "AWS FPGA: ERROR - Please install/enable Vivado." ; return 1; }
+vivado -version > /dev/null 2>&1 || { err_msg "Please install/enable Vivado."; return 1; }
 
-#Searching for Vivado version and comparing it with the list of supported versio
+#Searching for Vivado version and comparing it with the list of supported versions
 
 export VIVADO_VER=`vivado -version | grep Vivado | head -1`
 
-echo "AWS FPGA: Found $VIVADO_VER"
+info_msg "Found $VIVADO_VER"
 
 if grep -Fxq "$VIVADO_VER" $AWS_FPGA_REPO_DIR/hdk/supported_vivado_versions.txt
 then
-    echo "AWS FPGA: $VIVADO_VER is supported by this HDK release."
+    info_msg "$VIVADO_VER is supported by this HDK release."
 else
-    echo "AWS FPGA: ERROR - $VIVADO_VER is not supported by this HDK release."
+    err_msg "$VIVADO_VER is not supported by this HDK release."
     return 1
 fi
 
-echo "AWS FPGA: Vivado check succeeded"
+info_msg "Vivado check succeeded"
 
-echo "AWS FPGA: Setting up environment variables"
+info_msg "Setting up environment variables"
 
 # Clear environment variables
 unset HDK_DIR
@@ -54,45 +62,82 @@ export HDK_COMMON_DIR=$HDK_DIR/common
 
 # Point to the latest version of AWS shell
 export HDK_SHELL_DIR=$(readlink -f $HDK_COMMON_DIR/shell_stable)
+hdk_shell_version=$(readlink $HDK_COMMON_DIR/shell_stable)
 
 # The CL_DIR is where the actual Custom Logic design resides. The developer is expected to override this.
 # export CL_DIR=$HDK_DIR/cl/developer_designs
 
-echo "AWS FPGA: Done setting environment variables.";
+info_msg "Done setting environment variables.";
+
+# Download correct shell DCP
+info_msg "Using HDK shell version $hdk_shell_version"
+info_msg "Checking HDK shell's checkpoint version"
+hdk_shell_dir=$HDK_SHELL_DIR/build/checkpoints/from_aws
+hdk_shell=$hdk_shell_dir/SH_CL_BB_routed.dcp
+hdk_shell_s3_bucket=aws-fpga-hdk-resources
+# Download the sha1
+aws s3 cp s3://$hdk_shell_s3_bucket/hdk/$hdk_shell_version/build/checkpoints/from_aws/SH_CL_BB_routed.dcp.sha1 $hdk_shell.sha1 --only-show-errors || { err_msg "Failed to download HDK shell's checkpoint version."; return 2; }
+exp_sha1=$(cat $hdk_shell.sha1)
+info_msg "  latest   version=$exp_sha1"
+# If shell already downloaded check its sha1
+if [ -e $hdk_shell ]; then
+  act_sha1=$( sha1sum $hdk_shell | awk '{ print $1 }' )
+  info_msg "  existing version=$act_sha1"
+  if [[ $act_sha1 != $exp_sha1 ]]; then
+    info_msg "HDK shell's checkpoint version is incorrect"
+    info_msg "  Saving old checkpoint to $hdk_shell.back"
+    mv $hdk_shell $hdk_shell.back
+  fi
+else
+  info_msg "HDK shell's checkpoint hasn't been downloaded yet."
+fi
+if [ ! -e $hdk_shell ]; then
+  s3_hdk_shell=s3://$hdk_shell_s3_bucket/hdk/$hdk_shell_version/build/checkpoints/from_aws/SH_CL_BB_routed.dcp
+  info_msg "Downloading latest HDK shell checkpoint from $s3_hdk_shell"
+  aws s3 cp $s3_hdk_shell $hdk_shell --only-show-errors || { err_msg "HDK shell checkpoint download failed"; return 2; }
+fi
+# Check sha1
+act_sha1=$( sha1sum $hdk_shell | awk '{ print $1 }' )
+if [[ $act_sha1 != $exp_sha1 ]]; then
+  err_msg "Incorrect HDK shell checkpoint version:"
+  err_msg "  expected version=$exp_sha1"
+  err_msg "  actual   version=$act_sha1"
+  err_msg "  There may be an issue with the uploaded checkpoint."
+  return 2
+fi
+info_msg "HDK shell is up-to-date: $hdk_shell"
 
 # Create DDR and PCIe IP models and patch PCIe\
 models_dir=$HDK_COMMON_DIR/verif/models
 ddr4_model_dir=$models_dir/ddr4_model
 if [ ! -f $ddr4_model_dir/arch_defines.v ]; then
   ddr4_build_dir=$AWS_FPGA_REPO_DIR/ddr4_model_build
-  echo "AWS FPGA: DDR4 model files in "$ddr4_model_dir/" do NOT exist. Running model creation step.";
-  echo "AWS FPGA: Building in $ddr4_build_dir"
-  echo "AWS FPGA: This could take 5-10 minutes, please be patient!";
+  info_msg "DDR4 model files in "$ddr4_model_dir/" do NOT exist. Running model creation step.";
+  info_msg "Building in $ddr4_build_dir"
+  info_msg "This could take 5-10 minutes, please be patient!";
   return 2
   mkdir -p $ddr4_build_dir
   pushd $ddr4_build_dir &> /dev/null
   # Run init.sh then clean-up
   if ! $HDK_DIR/common/verif/scripts/init.sh $models_dir; then
-    echo "AWS FPGA: ERROR - DDR4 model build failed."
-    echo "AWS FPGA: Build dir=$ddr4_build_dir"
+    err_msg "DDR4 model build failed."
+    info_msg "Build dir=$ddr4_build_dir"
     popd &> /dev/null
     return 2
   fi
-  echo "AWS FPGA: Done with model creation step. Cleaning up temporary files.";
+  info_msg "Done with model creation step. Cleaning up temporary files.";
   popd &> /dev/null
   rm -rf $ddr4_build_dir
 else
-  echo "AWS FPGA: DDR4 model files exist in "$ddr4_model_dir/". Skipping model creation step.";
+  info_msg "DDR4 model files exist in "$ddr4_model_dir/". Skipping model creation step.";
 fi
-echo "AWS FPGA: Done with AWS HDK setup.";
+info_msg "Done with AWS HDK setup.";
 if [[ ":$CL_DIR" == ':' ]]; then
-  echo "AWS FPGA: ATTENTION: Don't forget to change the CL_DIR variable for the directory of your Custom Logic.";
+  info_msg "ATTENTION: Don't forget to change the CL_DIR variable for the directory of your Custom Logic.";
 else
-  echo "AWS FPGA: CL_DIR is $CL_DIR"
+  info_msg "CL_DIR is $CL_DIR"
   if [ ! -d $CL_DIR ]; then
-    echo "AWS FPGA: ERROR - CL_DIR doesn't exist. Set CL_DIR to a valid directory."
+    err_msg "CL_DIR doesn't exist. Set CL_DIR to a valid directory."
     unset CL_DIR
   fi
 fi
-
-
